requirements.txt
langchain
langchain_community
langchain-huggingface
langchain_chroma
python-dotenv
streamlit
replicate


